+++ make.py
@@ -2,13 +2,14 @@
 
 install: Poetry install
 build: Building docs, requirements.txt, setup.py, poetry build
+security: Run some basic security checks that are not run in vscode
 """
 import os
 import argparse
 
-HELP = "one of [install, build]"
+HELP = "one of [install, build, security]"
 
-parser = argparse.ArgumentParser()
+parser = argparse.ArgumentParser(description=__doc__)
 parser.add_argument("subcommand", action="store", help=HELP)
 args = parser.parse_args()
 
@@ -21,3 +22,5 @@ if args.subcommand == "build":
 elif args.subcommand == "install":
 	print("Poetry install")
 	os.system("poetry install")
+elif args.subcommand == "security":
+	os.system("poetry export -f requirements.txt | safety check --stdin")