+++ test/test.py
@@ -5,30 +5,37 @@ from pathlib import Path
 THISDIR = str(Path(__file__).resolve().parent)
 sys.path.insert(0, os.path.dirname(THISDIR))
 
-from stegstash import imagelsb, fileappend, soundlsb, homoglyphs, msoffice, odf
+from stegstash import (imagelsb, fileappend, soundlsb, homoglyphs, msoffice, odf, zerowidth)
 
 # Simple imagelsb
+print("\n# simple imagelsb")
 imagelsb.simpleEncode(THISDIR + "/originalImage.png",
 THISDIR + "/lsbSimpleEncode.png", "hello world from lsbSimpleEncode!")
 print(imagelsb.simpleDecode(THISDIR + "/lsbSimpleEncode.png"))
 
 # Hide image simplelsb
+print("\n# hide image simplelsb")
 imagelsb.simpleEncode(THISDIR + "/originalImage.png",
 THISDIR + "/lsbImageEncode.png", open(THISDIR + "/hideImage.png", "rb"))
 imagelsb.simpleDecode(THISDIR + "/lsbImageEncode.png", False,
 open(THISDIR + "/hideImageRecovered.png", "wb"))
 
 # imagelsb
+print("\n# imagelsb")
 imagelsb.encode(THISDIR + "/originalImage.png", THISDIR + "/lsbEncode.png",
 "hello world from lsbEncode!", "test", "pass")
 print(imagelsb.decode(THISDIR + "/lsbEncode.png", "test", "pass"))
 
-# imageappend
+# fileappend
+print("\n# fileappend")
 fileappend.encode(THISDIR + "/originalImage.png",
 THISDIR + "/appendEncode.png", "hello world from appendEncode!", "pass")
 print(fileappend.decode(THISDIR + "/appendEncode.png", "pass"))
+print(fileappend.detectSteg(THISDIR + "/originalImage.png"))
+print(fileappend.detectSteg(THISDIR + "/appendEncode.png"))
 
 # soundlsb
+print("\n# soundlsb")
 soundlsb.simpleEncode(THISDIR + "/originalSound.wav",
 THISDIR + "/simpleEncode.wav", "hello world from soundLsbSimpleEncode!")
 print(soundlsb.simpleDecode(THISDIR + "/simpleEncode.wav"))
@@ -37,6 +44,7 @@ soundlsb.encode(THISDIR + "/originalSound.wav", THISDIR + "/encode.wav",
 print(soundlsb.decode(THISDIR + "/encode.wav", "test", "pass"))
 
 # homoglyphs
+print("\n# homoglyphs")
 homoglyphs.simpleEncode(THISDIR + "/originalText.txt",
 THISDIR + "/simpleEncode.txt", "glyph")
 print(homoglyphs.simpleDecode(THISDIR + "/simpleEncode.txt"))
@@ -47,6 +55,7 @@ print(homoglyphs.detectSteg(THISDIR + "/originalText.txt"))
 print(homoglyphs.detectSteg(THISDIR + "/encode.txt"))
 
 # MsOffice
+print("\n# msoffice")
 msoffice.encodeComment(THISDIR + "/originalDoc.docx",
 THISDIR + "/encodeComment.docx", "hello world from encodeComment!")
 print(msoffice.decodeComment(THISDIR + "/encodeComment.docx"))
@@ -54,8 +63,11 @@ msoffice.encodeFile(THISDIR + "/originalDoc.docx",
 THISDIR + "/encodeFile.docx", open(THISDIR + "/hideImage.png", "rb"), password="pass")
 msoffice.decodeFile(THISDIR + "/encodeFile.docx", "pass",
 open(THISDIR + "/docxImageRecovered.png", "wb"))
+print(msoffice.detectSteg(THISDIR + "/originalDoc.docx"))
+print(msoffice.detectSteg(THISDIR + "/encodeFile.docx"))
 
 # odt
+print("\n# odt")
 odf.encodeComment(THISDIR + "/originalDoc.odt",
 THISDIR + "/encodeComment.odt", "hello world from encodeComment odt!")
 print(odf.decodeComment(THISDIR + "/encodeComment.odt"))
@@ -63,3 +75,16 @@ odf.encodeFile(THISDIR + "/originalDoc.odt",
 THISDIR + "/encodeFile.odt", open(THISDIR + "/hideImage.png", "rb"), password="pass")
 odf.decodeFile(THISDIR + "/encodeFile.odt", "pass",
 open(THISDIR + "/odtImageRecovered.png", "wb"))
+print(odf.detectSteg(THISDIR + "/originalDoc.odt"))
+print(odf.detectSteg(THISDIR + "/encodeFile.odt"))
+
+# zerowidth
+print("\n# zerowidth")
+zerowidth.simpleEncode(THISDIR + "/originalText.txt",
+THISDIR + "/simpleEncodeZW.txt", "zerowidth")
+print(zerowidth.simpleDecode(THISDIR + "/simpleEncodeZW.txt"))
+zerowidth.encode(THISDIR + "/originalText.txt",
+THISDIR + "/encodeZW.txt", "zerowidth", "test", "pass")
+print(zerowidth.decode(THISDIR + "/encodeZW.txt", "test", "pass"))
+print(zerowidth.detectSteg(THISDIR + "/originalText.txt"))
+print(zerowidth.detectSteg(THISDIR + "/simpleEncodeZW.txt"))