+++ CHANGELOG.md
@@ -2,14 +2,17 @@
 All major and minor version changes will be documented in this file. Details of
 patch-level version changes can be found in [commit messages](../../commits/master).
 
-## 2020.1 - 2020/05/22
+## 2020.1 - 2020/06/02
 - Hide binary in sound and image files (switched from text to binary for this)
 - Can now hide whole images inside other images!
-- Use map and otp lsb encodings for sound
+- Use map and OTP LSB encodings for sound
 - Image mode is preserved
 - Removed CLI
 - Can hide data in Microsoft Office files
 - Can hide data in OpenOffice/ LibreOffice files
+- Optionally encrypt data and use a mapping for homoglyphs
+- Hide using zero width chars
+- Detect some forms of steganography with varying success
 
 ## 2020 - 2020/05/22
 - First release